+++ src/firebaseApi.ts
@@ -4,8 +4,8 @@
 import { SnapError, SnapErrorCode } from './error';
 
 const baseUrl = process.env.IS_PRODUCTION
-	? 'https://cloudfuntion-prod.silencelaboratories.com'
-	: 'https://cloudfuntion-stage.silencelaboratories.com';
+	? 'https://us-central1-mobile-wallet-mm-snap.cloudfunctions.net'
+	: 'https://us-central1-mobile-wallet-mm-snap-staging.cloudfunctions.net';
 
 interface Response {
 	response: any;