+++ src/index.ts
@@ -9,7 +9,7 @@ import {
 import * as sdk from './snap/sdk';
 import { deleteStorage, getSilentShareStorage } from './snap/storage';
 import { isPaired } from './snap/sdk';
-import { panel, text, heading, divider } from '@metamask/snaps-ui';
+import { panel, text, heading, divider } from '@metamask/snaps-sdk';
 import { SnapError, SnapErrorCode } from './error';
 import { handleKeyringRequest } from '@metamask/keyring-api';
 import { SimpleKeyring } from './snap/keyring';
@@ -172,7 +172,6 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
 		 */
 		case InternalMethod.TssRunKeygen:
 			const keygenRes = await sdk.runKeygen();
-			await sdk.runBackup();
 			return {
 				address:
 					'0x' +
@@ -180,6 +179,19 @@ export const onRpcRequest: OnRpcRequestHandler = async ({
 						Buffer.from(keygenRes.distributedKey.publicKey, 'hex'),
 					).toString('hex'),
 			};
+
+		/**
+		 * tss_runBackup
+		 * @abstract This methods update the backup and phone can listen to the backup
+		 *
+		 * @throws
+		 * 1. NotPaired, if snap is not paired yet.
+		 * 2. FirebaseError, when error occurs on server side, message will contains info,
+		 * 3. UnknownError, when something unknown error occurs
+		 * 4. BackupFailed, when distributed key not found
+		 */
+		case InternalMethod.TssRunBackup:
+			await sdk.runBackup();
 
 		/**
 		 * tss_snapVersion