+++ src/snap/utils.ts
@@ -4,7 +4,6 @@
 import { SnapError, SnapErrorCode } from '../error';
 import { JsonTx } from '@ethereumjs/tx';
 import type { Json } from '@metamask/utils';
-import { randBytes } from '@silencelaboratories/ecdsa-tss';
 import { pubToAddress } from 'ethereumjs-util';
 import _sodium from 'libsodium-wrappers';
 import { DistributedKey } from '../types';